+++ src/main/java/io/drakon/pulsar/pulse/PulseMeta.java
@@ -1,28 +1,25 @@
 package io.drakon.pulsar.pulse;
 
-import com.google.common.eventbus.EventBus;
-import net.minecraftforge.fml.common.Loader;
-
-import io.drakon.pulsar.internal.BusExceptionHandler;
+import javax.annotation.Nullable;
+import javax.annotation.ParametersAreNonnullByDefault;
 
 /**
  * Metadata wrapper for parsed @Pulse metadata.
  *
  * @author Arkan <arkan@drakon.io>
  */
+@ParametersAreNonnullByDefault
 public class PulseMeta {
 
     private String id, description;
     private boolean forced, enabled, defaultEnabled;
-    public final EventBus bus;
 
-    public PulseMeta(String id, String description, boolean forced, boolean enabled, boolean defaultEnabled) {
+    public PulseMeta(String id, @Nullable String description, boolean forced, boolean enabled, boolean defaultEnabled) {
         this.id = id;
         this.description = description;
         this.forced = forced;
         this.enabled = enabled;
         this.defaultEnabled = defaultEnabled;
-        this.bus = new EventBus(new BusExceptionHandler(Loader.instance().activeModContainer().getModId(), id));
     }
 
     public String getId() {